structure createBiggerBOM functionality within DB && add layer functionality to 1,000/10,000 BOMS.
