+++ src/db/createBiggerBOM/index.css
@@ -1,0 +1,15 @@
+
+body {
+    margin: 40px;
+}
+
+#root {
+    display: flex;
+    align-items: center;
+}
+
+input {
+    cursor: pointer;
+    margin: 0 20px 10px 0;
+    padding: 4px 10px;
+}
