+++ src/db/createBiggerBOM/php/grabList.php
@@ -1,0 +1,14 @@
+
+<?php
+    function grabList( $result ) { // convert an object into an array
+        $list = [];
+    
+        if( isset( $result ) ) {
+          foreach( $result as $key => $val ) {
+            array_push( $list, $val );
+          }
+        }
+    
+        return $list;
+    }
+?>
