+++ src/db/createBiggerBOM/php/json.php
@@ -1,0 +1,8 @@
+
+<?php
+    function json( $BOM ) { // convert array to json
+        $json = json_encode( $BOM, JSON_PRETTY_PRINT );
+
+        return $json;
+    }
+?>
