+++ src/db/db_credentials.php
@@ -3,5 +3,5 @@
     DEFINE('DB_SERVER', 'localhost');
     DEFINE('DB_NAME', 'sbom_db');
     DEFINE('DB_USER', 'root');
-    DEFINE('DB_PASS', '');
+    DEFINE('DB_PASS', 'password');
 ?>
